% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{ChangeWorklist}
\alias{ChangeWorklist}
\title{ChangeWorklist}
\usage{
ChangeWorklist(date = gsub("-", "", as.character(Sys.Date())))
}
\arguments{
\item{date}{Date you want to change in you worklist. Default is today.}
}
\value{
Return a new worklist.
}
\description{
Change date in worklist from GetWorklist.
}
\examples{
\donttest{
#demo data
batch.design <- paste("A",c(1:100), sep = "")
##create a folder for demo
dir.create("Demo")
setwd("Demo")
write.csv(batch.design, "batch.design.csv", row.names = FALSE)
#get worklist
GetWorklist()
##remove other information
file.remove(dir()[dir()!="worklist POS.csv"])
#run ChangeWorklist
ChangeWorklist()
}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
