\name{data}
\alias{data}
\docType{data}
\title{
Metabolomics dataset for demo.
}
\description{
Metabolomics dataset has two batches. 1401 features and 610 samples.
}
\usage{data(data)}
\format{
  \describe{
    \item{\code{name}}{Feature name}
    \item{\code{isotope}}{Isotope annotation}
    \item{\code{mz}}{Mass to charge ratio}
    \item{\code{rt}}{Retention time}
  }
}

\value{
Metabolomics data.
}

\examples{
data(data)
}
\keyword{datasets}
