% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataIntegration.R
\name{DataIntegration}
\alias{DataIntegration}
\title{DataIntegration}
\usage{
DataIntegration(MetFlowData, method = c("qc.mean", "subject.mean"))
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{method}{Which method do you want to use?
subject.mean or qc.mean, default is qc.mean.}
}
\value{
Return a MetFlowData which has been integrated.
}
\description{
Integrate different batch datasets together.
}
\details{
Data integration is a necessary step for multiple batch dataset.
\href{https://www.readcube.com/library/fe13374b-5bc9-4c61-9b7f-6a354690947e:abe41368-d08d-4806-871f-3aa035d21743}{Dunn's}
method has been used in this function.
}
\examples{
\donttest{
#load the demo data
data(data, package = "metflowR")
data(sample.information, package = "metflowR")

##create a folder for demo
dir.create("demo")
setwd("demo")

# export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample.information, "sample.information.csv", row.names = FALSE)

# metflowR process
metflowR(#ImportData para
data = "data.csv",
sample.information = "sample.information.csv",
polarity = "positive",
#DataNormalization
method = "svr",
threads = 2)
## run DataIntegration
new.met.data <- DataIntegration(MetFlowData = met.data.after.pre)
}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
