% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normationFunction.R
\name{DataNormalization}
\alias{DataNormalization}
\title{DataNormalization}
\usage{
DataNormalization(MetFlowData, path = ".", method = "svr",
  dimension1 = TRUE, optimization = TRUE, begin = 0.5, end = 1,
  step = 0.2, multiple = 5, threads = 2, rerun.loess = TRUE,
  rerun.svr = TRUE, peakplot = TRUE, datastyle = "tof", user = "other")
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{path}{Workdirectory.}

\item{method}{Normalization method, mean, median, total svr or loess,
default is svr. Please see the details.}

\item{dimension1}{Remain dimension or not. Default is TRUE.}

\item{optimization}{Optimize span and degree or not.}

\item{begin}{Begin of span.}

\item{end}{End of span.}

\item{step}{Step of span.}

\item{multiple}{See ?SXTsvrNor.}

\item{threads}{Thread number.}

\item{rerun.loess}{Rerun loess or not.}

\item{rerun.svr}{Rerun SVR or not.}

\item{peakplot}{Peak plot or not.}

\item{datastyle}{Default is "tof".}

\item{user}{Default is "other".}
}
\value{
The normalization results can be got from help of
\code{\link{MetNormalizer}}.
}
\description{
Normalize dataset.
}
\details{
Data normalization is a useful method to reduce unwanted variances
in metabolomics data. The most used normalization methods is
\href{https://www.readcube.com/library/fe13374b-5bc9-4c61-9b7f-6a354690947e:2303e537-2864-4df5-91f0-c3ae731711af}{sample scale method}
(median, mean and total intensity). In large
scale metabolomics, QC based normalization methods are more useful.
\href{https://www.readcube.com/library/fe13374b-5bc9-4c61-9b7f-6a354690947e:abe41368-d08d-4806-871f-3aa035d21743}{LOESS}
and \href{https://www.readcube.com/library/fe13374b-5bc9-4c61-9b7f-6a354690947e:2303e537-2864-4df5-91f0-c3ae731711af}{SVR}
normalization are two most used normalization
methods.
}
\examples{
\donttest{
#load the demo data
data(data, package = "metflowR")
data(sample.information, package = "metflowR")

##create a folder for demo
dir.create("demo")
setwd("demo")

# export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample.information, "sample.information.csv", row.names = FALSE)

# metflowR process
metflowR(#ImportData para
data = "data.csv",
sample.information = "sample.information.csv",
polarity = "positive",
#DataNormalization
method = "svr",
threads = 2)
## run
new.met.data <- DataNormalization(MetFlowData = met.data.after.pre)
}
}
\seealso{
\code{\link{MetNormalizer}}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
