% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsForMetflowR.R
\name{ExportData}
\alias{ExportData}
\title{ExportData}
\usage{
ExportData(MetFlowData, data.name = "data_new",
  subject.info.name = "subject.info", qc.info.name = "qc.info",
  path = ".")
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{data.name}{The name of the data you want to output. Default is
"data_new".}

\item{subject.info.name}{The name for subject information you want to
output. Default is "subject.info".}

\item{qc.info.name}{The name for QC information you want to output.
Default is "qc.info".}

\item{path}{Work directory.}
}
\value{
Write csv data.
}
\description{
Export MetFlowData as csv.
}
\examples{
#load the demo data
data(met.data, package = "metflowR")
##create a folder for demo
dir.create("demo")
setwd("demo")
## run
ExportData(MetFlowData = met.data, path = "Demo for ExportData")
}
\seealso{
\code{\link{ImportData}}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
