% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{GetWorklist}
\alias{GetWorklist}
\title{GetWorklist}
\usage{
GetWorklist(x = NULL, instrument = "Agilent", name = "worklist",
  randommethod = "no", samplenumber = NULL, replication = 1, QCstep = 8,
  conditionQCnumber = 10, testmixstep = 0, injectionfrom = 1,
  user = "other", dir = "D:\\\\MassHunter\\\\Data\\\\SXT\\\\")
}
\arguments{
\item{x}{batch.design file.}

\item{instrument}{Which instrument you use?
"Agilent" or "AB", default is "Agilent".}

\item{name}{The name of worklist.}

\item{randommethod}{Which random method you want to use?
"no", "position" or "injection". Default is "no".}

\item{samplenumber}{Sample number.}

\item{replication}{Replication times.}

\item{QCstep}{QC step.}

\item{conditionQCnumber}{Condition QC number.}

\item{testmixstep}{Test mixture step.}

\item{injectionfrom}{Injection order from which? Default is 1.}

\item{user}{Default is "other".}

\item{dir}{Directory.}
}
\value{
New worklist.
}
\description{
Generate Worklist for data acquisition.
}
\details{
Generate Worklist for data acquisition.
}
\examples{
\donttest{
#demo data
batch.design <- paste("A",c(1:100), sep = "")
##create a folder for demo
dir.create("Demo")
setwd("Demo")
write.csv(batch.design, "batch.design.csv", row.names = FALSE)

#run ChangeWorklist
GetWorklist()
}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
