% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{HeatMap}
\alias{HeatMap}
\title{HeatMap}
\usage{
HeatMap(MetFlowData, log.scale = FALSE, color = c("palegreen",
  "firebrick1", "royalblue", "yellow", "black", "cyan", "gray48"),
  variable = "all", Group = c("control", "case"),
  scale.method = "auto", show_rownames = FALSE,
  show_colnames = FALSE, path = ".", width = 7, height = 7,
  border_color = NA, fontsize_row = 10, cluster_rows = TRUE,
  cluster_cols = TRUE, clustering_method = "ward.D", ...)
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{log.scale}{log transformation or not.}

\item{color}{Color list for sample group.}

\item{variable}{"all" or "marker" for heatmap.}

\item{Group}{group for heatmap.}

\item{scale.method}{scale method.}

\item{show_rownames}{Default is FALSE.}

\item{show_colnames}{Default is FALSE.}

\item{path}{Work directory.}

\item{width}{Plot width}

\item{height}{Plot height.}

\item{border_color}{Default is NA.}

\item{fontsize_row}{Default is 10.}

\item{cluster_rows}{Default is TRUE.}

\item{cluster_cols}{Default is TURE.}

\item{clustering_method}{Default is"ward.D",}

\item{...}{other parameters for pheatmap.}
}
\value{
A heatmap plot.
}
\description{
Heat map
}
\examples{
\donttest{
#load the demo data
data(data, package = "metflowR")
data(sample.information, package = "metflowR")

##create a folder for demo
dir.create("demo")
setwd("demo")

# export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample.information, "sample.information.csv", row.names = FALSE)

# metflowR process
metflowR(#ImportData para
data = "data.csv",
sample.information = "sample.information.csv",
polarity = "positive",
#DataNormalization
method = "svr",
threads = 2)
## run
HeatMap(MetFlowData = met.data.after.pre,
       path = "Demo for HeatMap",
       Group = c("0", "1"),
       path = "Demo for HeatMap")
       }
}
\seealso{
\code{\link[pheatmap]{pheatmap}}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
