\name{hmdbdatabase}
\alias{hmdbdatabase}
\docType{data}
\title{
HMDB database.
}
\description{
HMDB database.
}
\usage{data(hmdbdatabase)}
\format{
  \describe{
    \item{\code{HMDB.ID}}{HMDB ID.}
    \item{\code{name}}{Metabolite name.}
    \item{\code{Mass}}{Accurate mass.}
    \item{\code{formula}}{Chemical formual.}
  }
}

\value{
HMDB database.
}

\examples{
data(hmdbdatabase)
}
\keyword{datasets}