% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsForMetCleaning.R
\name{ImportData}
\alias{ImportData}
\title{ImportData}
\usage{
ImportData(data = "data.csv", sample.information = "sample.information.csv",
  polarity = c("positive", "negative", "none"), hasIS = c("yes", "no"),
  hasQC = c("yes", "no"), peak.identification = c("yes", "no"),
  path = ".")
}
\arguments{
\item{data}{Data name for analysis. Default is "data.csv". Data are csv
format from XCMS, MZmine or other software.
 Please see the demo data in example.}

\item{sample.information}{Sample information name for analysis. Default is
"sample.information.csv". Column 1 is sample.name, column 2 is
injection.order, column 3 is class ("Subject" or "QC"), column 4 is
batch information, column 5 is group ("control" or "case"), other columns
are information for sample. Please see demo data in example.}

\item{polarity}{The polarity of data, "positive", "negative"" or "none"",
default is positive.}

\item{hasIS}{The data has IS or not? Default is "no".}

\item{hasQC}{The data has QC samples or not? Default is "yes".}

\item{peak.identification}{The data has identification result or not?
Default is "no".}

\item{path}{Work directory.}
}
\value{
Return a standard MetProcesser dataset.
}
\description{
Import data and retrun a standard MetProcessor data.
}
\examples{
#load the demo data
data(data, package = "metflowR")
data(sample.information, package = "metflowR")
# export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample.information, "sample.information.csv", row.names = FALSE)
#Import data
met.data <- ImportData()
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
