% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerFunction.R
\name{MarkerSelection}
\alias{MarkerSelection}
\title{MarkerSelection}
\usage{
MarkerSelection(MetFlowData, foldchange = "foldchange", p = "p.correct",
  vip = "vip", foldchange.cutoff = c(4/3, 3/4), p.cutoff = 0.05,
  vip.cutoff = 1, path = ".")
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{foldchange}{which foldchange information your want to use?}

\item{p}{which p value information your want to use?}

\item{vip}{VIP.}

\item{foldchange.cutoff}{The cutoff value of fold change.}

\item{p.cutoff}{The cutoff of p value.}

\item{vip.cutoff}{The cutoff of VIP value.}

\item{path}{The directory you want to write results.}
}
\value{
MetFlowData: MetFlowData which has been added is.
marker information into tags.

maker.information.csv: The marker information.
}
\description{
Select markers using fold change and p values.
}
\details{
Select markers using fold change and p values.
}
\examples{
\donttest{
#load the demo data
data(data, package = "metflowR")
data(sample.information, package = "metflowR")

##create a folder for demo
dir.create("demo")
setwd("demo")

# export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample.information, "sample.information.csv", row.names = FALSE)

# metflowR process
metflowR(#ImportData para
data = "data.csv",
sample.information = "sample.information.csv",
polarity = "positive",
#DataNormalization
method = "svr",
threads = 2)
# FoldChange
met.data <- FoldChange(MetFlowData = met.data.after.pre,
                      to = c("1", "0"))
##PLSanalysis
PLSanalysis(met.data,
           plsmethod = "plsr",
           path = "Demo for PLSanalysis")

load("Demo for PLSanalysis/vip")
vip <- apply(vip, 2, mean)
##VIP
tags <- met.data[["tags"]]
tags <- data.frame(tags, vip)
met.data[["tags"]] <- tags
##UnivariateTest
met.data <- UnivariateTest(met.data)
## run
new.met.data <- MarkerSelection(met.data)
}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
