% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerFunction.R
\name{MarkerShow}
\alias{MarkerShow}
\title{MarkerShow}
\usage{
MarkerShow(MetFlowData, beeswarm = TRUE, path = ".")
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{beeswarm}{Do you want draw beeswarm on the boxplot? Deafult is TRUE.}

\item{path}{The directory you want to write results.}
}
\value{
Box plot of markers.
}
\description{
Draw boxplot for markers.
}
\examples{
\donttest{
#load the demo data
data(data, package = "metflowR")
data(sample.information, package = "metflowR")

##create a folder for demo
dir.create("demo")
setwd("demo")

# export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample.information, "sample.information.csv", row.names = FALSE)

# metflowR process
metflowR(#ImportData para
data = "data.csv",
sample.information = "sample.information.csv",
polarity = "positive",
#DataNormalization
method = "svr",
threads = 2)
# FoldChange
met.data <- FoldChange(MetFlowData = met.data.after.pre,
                      to = c("1", "0"))
##PLSanalysis
PLSanalysis(met.data,
           plsmethod = "plsr",
           path = "Demo for PLSanalysis")

load("Demo for PLSanalysis/vip")
vip <- apply(vip, 2, mean)
##VIP
tags <- met.data[["tags"]]
tags <- data.frame(tags, vip)
met.data[["tags"]] <- tags
##UnivariateTest
met.data <- UnivariateTest(met.data)
## run
new.met.data <- MarkerSelection(met.data)
run
MarkerShow(new.met.data, path = "Demo for MarkerShow")
}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
