% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetStat.R
\name{MetStat}
\alias{MetStat}
\title{MetStat}
\usage{
MetStat(MetFlowData, new.group = TRUE, rsd.cutoff = 30, log.scale = FALSE,
  QC = TRUE, scale.method = "auto", xlim1 = NULL, ylim1 = NULL,
  plsmethod = "plsr", fc = TRUE, to = c("case", "control"),
  ratio = "median", test.method = "t", adjust.method = "fdr",
  foldchange = "foldchange", p = "p", vip = "vip",
  foldchange.cutoff = c(4/3, 3/4), p.cutoff = 0.05, vip.cutoff = 0,
  x = "foldchange", y = "p", z = "vip", col = c("black", "firebrick1"),
  variable = "all", Group = c("control", "case"), beeswarm = TRUE,
  pls.analysis = TRUE, pca.analysis = TRUE, uni.test = TRUE, path = ".")
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{new.group}{New group information or not?}

\item{rsd.cutoff}{RSD cutoff.}

\item{log.scale}{Data transformation method. Default is FALSE.}

\item{QC}{Use qc data for PCA analyis or not? Default is FALSE.}

\item{scale.method}{Which scale methd you want to use? "auto" or "pareto",
defaulit is "auto".}

\item{xlim1}{x axis limitation. Default is NULL.}

\item{ylim1}{y axis limitation. Default is NULL.}

\item{plsmethod}{Default is "plsreg".}

\item{fc}{Default is TRUE.}

\item{to}{Which the ratio you want? default is case/control.}

\item{ratio}{Which ratio you want to use to calculate fold
change. median ot mean.}

\item{test.method}{Which test you want to use? "t" means
stutent t test and "wilcox" mean wilcoxon test.}

\item{adjust.method}{p value correction method. See p.adjust function.}

\item{foldchange}{which foldchange information your want to use?}

\item{p}{which p value information your want to use?}

\item{vip}{VIP.}

\item{foldchange.cutoff}{The cutoff value of fold change.}

\item{p.cutoff}{The cutoff of p value.}

\item{vip.cutoff}{The cutoff of VIP value.}

\item{x}{x axis factor.}

\item{y}{y axis factor.}

\item{z}{z axis factor.}

\item{col}{Colour for markers and non-markers.}

\item{variable}{"all" or "marker" for heatmap.}

\item{Group}{group for heatmap.}

\item{beeswarm}{Do you want draw beeswarm on the boxplot? Deafult is TRUE.}

\item{pls.analysis}{Do PLS or not.}

\item{pca.analysis}{Do PCA or not.}

\item{uni.test}{Do univariate test or not.}

\item{path}{The directory you want to write results.}
}
\value{
All the results can be got form other functions and instruction.
}
\description{
A whole work flow for high throughput MS based metabolomics
data statistical analysis.
}
\details{
The manual of MetCleaning can be found in \href{https://github.com/jaspershen/MetCleaning/blob/master/vignettes/MetCleaning.pdf}{github}
or in \href{https://www.readcube.com/library/fe13374b-5bc9-4c61-9b7f-6a354690947e:d8b5fcff-c725-4689-a97d-7ff106322fb6}{my library}.
}
\examples{
\donttest{
#' #load the demo data
data(data, package = "MetCleaning")
data(sample.information, package = "MetCleaning")

##create a folder for MetCleaning demo
dir.create("Demo for MetCleaning")
setwd("Demo for MetCleaning")

# export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample.information, "sample.information.csv", row.names = FALSE)

#run MetCleaning
MetCleaning(#ImportData para
data = "data.csv",
sample.information = "sample.information.csv",
polarity = "positive",
#DataNormalization
method = "svr",
threads = 2)

## load the demo data
data(new.group, package = "MetCleaning")
load("met.data.after.pre")
##create a folder for MetStat demo
dir.create("Demo for MetStat")
setwd("Demo for MetStat")
## export the demo data as csv
write.csv(new.group, "new.group.csv", row.names = FALSE)

## run MetStat
MetStat(MetFlowData = met.data.after.pre,
new.group = TRUE,
Group = c("0", "1"),
to = c("1", "0"))
}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
