% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVZFunction.R
\name{MVimputation}
\alias{MVimputation}
\title{MVimputation}
\usage{
MVimputation(MetFlowData, imputation.method = "knn", k = 10, rowmax = 0.5,
  colmax = 0.8, maxp = 1500, rng.seed = 362436069, maxiter = 10,
  ntree = 100, decreasing = FALSE, replace = TRUE, classwt = NULL,
  cutoff = NULL, strata = NULL, sampsize = NULL, nodesize = NULL,
  maxnodes = NULL, xtrue = NA, parallelize = "no", nPcs = 2,
  maxSteps = 100, threshold = 1e-04)
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{imputation.method}{Which imputation method you want to use? It
contains "knn", "rf" (missForest), "mean", "median", "zero", "minium",
"bpca" (BPCA), "svd" (SVD) and "ppca" (PPCA). Default is "knn".
The detial of
this method can be find in detail and reference paperes.}

\item{k}{See ?impute.knn}

\item{rowmax}{See ?impute.knn}

\item{colmax}{See ?impute.knn}

\item{maxp}{See ?impute.knn}

\item{rng.seed}{See ?impute.knn}

\item{maxiter}{See ?missForest}

\item{ntree}{See ?missForest}

\item{decreasing}{See ?missForest}

\item{replace}{See ?missForest}

\item{classwt}{See ?missForest}

\item{cutoff}{See ?missForest}

\item{strata}{See ?missForest}

\item{sampsize}{See ?missForest}

\item{nodesize}{See ?missForest}

\item{maxnodes}{See ?missForest}

\item{xtrue}{See ?missForest}

\item{parallelize}{See ?missForest}

\item{nPcs}{See ?bpca}

\item{maxSteps}{See ?bpca}

\item{threshold}{See ?bpca}
}
\value{
Return a MetFlowData whose MVs have been imputated.
}
\description{
Impute MV in data.
}
\examples{
\donttest{
#load the demo data
data(data, package = "MetCleaning")
data(sample.information, package = "MetCleaning")

##create a folder for demo
dir.create("demo")
setwd("demo")

# export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample.information, "sample.information.csv", row.names = FALSE)
#Import data
met.data <- ImportData(data = "data.csv",
                      sample.information = "sample.information.csv",
                      polarity = "positive")
#MV filtering
met.data <- MZfilter(MetFlowData = met.data,
                    obs.per.cutoff = 0.5,
                    var.per.cutoff = 0.5,
                    what = "mv",
                    path = "Demo for MV filter")
run
new.met.data <- MVimputation(met.data, rowmax = 0.9, colmax = 0.9)
}
}
\references{
The MV imputation in metabolomics data can see in
\href{https://www.readcube.com/library/fe13374b-5bc9-4c61-9b7f-6a354690947e:c9d05d0f-e945-43d0-bb4a-50ea0f90338e}{Guida's} paper.
}
\seealso{
The MV imputation methods can see in
\code{\link[impute]{impute.knn}}, \code{\link[missForest]{missForest}},
\code{\link[pcaMethods]{bpca}}, \code{\link[pcaMethods]{ppca}} and
\code{\link[pcaMethods]{svdImpute}}.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
