% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVZFunction.R
\name{MZfilter}
\alias{MZfilter}
\title{MZfilter}
\usage{
MZfilter(MetFlowData, obs.per.cutoff = 0.5, var.per.cutoff = 0.5,
  what = "mv", path = ".")
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{obs.per.cutoff}{The MV/zero ratio cutoff value of samples,
default is 0.5. It means that for a sample, non-MV ratio/zero must be larger
than obs.per.cutoff.}

\item{var.per.cutoff}{The MV/zero ratio cutoff value of features,
default is 0.5. It means that for a feature, non-MV/zero ratio must be
larger than obs.per.cutoff.}

\item{what}{Filter missing values ("mv") or zero values ("zero")?}

\item{path}{Work directory.}
}
\value{
Return a MetFlowData which has been filtered according MV/zero ratio.
}
\description{
Filter feature and samples according to MV/zero ratio.
}
\examples{
\donttest{
#load the demo data
data(data, package = "MetCleaning")
data(sample.information, package = "MetCleaning")

##create a folder for demo
dir.create("demo")
setwd("demo")

# export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample.information, "sample.information.csv", row.names = FALSE)
#Import data
met.data <- ImportData(data = "data.csv",
                      sample.information = "sample.information.csv",
                      polarity = "positive")
#MV filtering
new.met.data <- MZfilter(MetFlowData = met.data,
                        obs.per.cutoff = 0.5,
                        var.per.cutoff = 0.5,
                        what = "mv",
                        path = "Demo for MV filter")
                        }
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
