% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVZFunction.R
\name{MZoverview}
\alias{MZoverview}
\title{MZOverview}
\usage{
MZoverview(MetFlowData, path = ".", what = "mv")
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{path}{Work Directory.}

\item{what}{Missing value ("mv") or zero values ("zero").}
}
\value{
Batch i feature MV/zero ratio.csv: Batch i feature MV/zero ratio
distribution.

Batch i QC feature MV/zero ratio.pdf: Batch i QC feature MV.zero
ratio distribution.

Batch i Subject feature MV/zero ratio.pdf: Batch i Subject feature
 MV/zero ratio distribution.

Sample MV/zero distribution.csv: Sample MV/zero ratio distribution.
}
\description{
Evaluate the missing values or zero value information
in MetFlowData.
}
\examples{
\donttest{
#load the demo data
data(data, package = "MetCleaning")
data(sample.information, package = "MetCleaning")

##create a folder for demo
dir.create("demo")
setwd("demo")

# export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample.information, "sample.information.csv", row.names = FALSE)
#Import data
met.data <- ImportData(data = "data.csv",
                      sample.information = "sample.information.csv",
                      polarity = "positive")
#run
MZoverview(MetFlowData = met.data,
                    what = "mv",
                    path = "Demo for MZoverview")
                    }
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
