\name{new.group}
\alias{new.group}
\docType{data}
\title{
New group information.
}
\description{
New group information.
}
\usage{data("new.group")}
\format{
  A data frame with 606 observations on the following 5 variables.
  \describe{
    \item{\code{sample.name}}{Sample name}
    \item{\code{injection.order}}{Injection order}
    \item{\code{class}}{Class information}
    \item{\code{batch}}{Batch information}
    \item{\code{group}}{New group information}
  }
}

\value{
New group inforamtion.
}

\examples{
data(new.group)
}
\keyword{datasets}
