% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsFunction.R
\name{PLSanalysis}
\alias{PLSanalysis}
\title{PLSanalysis}
\usage{
PLSanalysis(MetFlowData, log.scale = 10, scalemethod = "auto",
  plsmethod = "plsreg", path = ".", width = 7, height = 7,
  text = FALSE, ellipse = TRUE, color = c("palegreen", "royalblue",
  "firebrick1", "yellow", "black", "cyan", "gray48"), shape = c(17, 19, 15,
  18, 2, 8, 11), cexa = 1, xlim1 = NULL, ylim1 = NULL)
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{log.scale}{Which transformation methd you want to use? 2, 10  or "e",
defaulit is FALSE mean don't transformation.}

\item{scalemethod}{Whihch scale methd you want to use? "auto" or "pareto",
defaulit is "auto".}

\item{plsmethod}{Default is "plsreg".}

\item{path}{Work directory.}

\item{width}{width of plot.}

\item{height}{height of plot.}

\item{text}{Add text in PCA score plot or not? Deefault is FALSE.}

\item{ellipse}{Add ellipse in PCA score plot or not? Deefault is TRUE.}

\item{color}{Color for different class.}

\item{shape}{Shape for different class.}

\item{cexa}{cex.}

\item{xlim1}{x axis limitation. Default is NULL.}

\item{ylim1}{y axis limitation. Default is NULL.}
}
\value{
PLS score plot: PLS score plot.

permutation test plot: Permutation test plot.
}
\description{
PLS analysis for MetFlowData.
}
\examples{
\donttest{
#load the demo data
data(data, package = "metflowR")
data(sample.information, package = "metflowR")

##create a folder for demo
dir.create("demo")
setwd("demo")

# export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample.information, "sample.information.csv", row.names = FALSE)

# metflowR process
metflowR(#ImportData para
data = "data.csv",
sample.information = "sample.information.csv",
polarity = "positive",
#DataNormalization
method = "svr",
threads = 2)
# FoldChange
met.data <- FoldChange(MetFlowData = met.data.after.pre)
##PLSanalysis
PLSanalysis(met.data,
           plsmethod = "plsr",
           path = "Demo for PLSanalysis")
}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
