% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsForMetflowR.R
\name{ReChangeGroup}
\alias{ReChangeGroup}
\title{ReChangeGroup}
\usage{
ReChangeGroup(MetFlowData, new.group = "new.group.csv")
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{new.group}{New Group information name. Default is new.group.csv.
It can be use the sample.information which is only changed the group column.}
}
\value{
Return a standard MetProcesser data which is changed the
group informatio.
}
\description{
Change the group information in MetFlowData for statistical
analysis.
}
\examples{
\donttest{
#load the demo data
data(data, package = "metflowR")
data(sample.information, package = "metflowR")
data(new.group, package = "metflowR")

##create a folder for demo
dir.create("demo")
setwd("demo")

# export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample.information, "sample.information.csv", row.names = FALSE)
write.csv(new.group, "new.group.csv", row.names = FALSE)

# metflowR process
metflowR(#ImportData para
data = "data.csv",
sample.information = "sample.information.csv",
polarity = "positive",
#DataNormalization
method = "svr",
threads = 2)

#run
new.met.data <- ReChangeGroup(met.data.after.pre)
}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
