% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsForMetCleaning.R
\name{RSDoverview}
\alias{RSDoverview}
\title{RSDoverview}
\usage{
RSDoverview(MetFlowData.before, MetFlowData.after, path = ".")
}
\arguments{
\item{MetFlowData.before}{MetFlowData.before.}

\item{MetFlowData.after}{MetFlowData.after.}

\item{path}{Work directory.}
}
\value{
RSD comparation.
}
\description{
Evaluate the RSD information before and after processing.
}
\examples{
#load the demo data
data(met.data, package = "metflowR")
##create a folder for demo
dir.create("demo")
setwd("demo")
RSDoverview(met.data, met.data, path = "Demo for RSDoverview")
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
