\name{sample.information}
\alias{sample.information}
\docType{data}
\title{
Sample information
}
\description{
Sample information
}

\value{
Sample information.
}

\usage{data("sample.information")}
\format{
  A data frame with 606 observations on the following 5 variables.
  \describe{
    \item{\code{sample.name}}{Sample name}
    \item{\code{injection.order}}{Injection order}
    \item{\code{class}}{Class information}
    \item{\code{batch}}{Batch information}
    \item{\code{group}}{Group information}
  }
}

\examples{
data(sample.information)
}
\keyword{datasets}
