% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintMethodsForMyClass.R
\docType{methods}
\name{show,MetFlowData-method}
\alias{show,MetFlowData-method}
\title{Print.MetFlowData}
\usage{
\S4method{show}{MetFlowData}(object)
}
\arguments{
\item{object}{MetFlowData.}
}
\value{
Print some information of class object in your screen.
}
\description{
Print method for class in MetProcesser.
}
\seealso{
\code{\link{print}}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
