% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsForMetCleaning.R
\name{SplitBatch}
\alias{SplitBatch}
\title{SplitBatch}
\usage{
SplitBatch(MetFlowData)
}
\arguments{
\item{MetFlowData}{MetFlowData.}
}
\value{
Return a data (list), subject, qc, subject.info and qc.info.
}
\description{
Split MetFlowData accoding to different batch.
}
\examples{
#load the demo data
data(met.data, package = "metflowR")
##create a folder for demo
dir.create("demo")
setwd("demo")
new.data <- SplitBatch(met.data)
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
