% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierFunction.R
\name{SubjectOutlierFilter}
\alias{SubjectOutlierFilter}
\title{SubjectOutlierFilter}
\usage{
SubjectOutlierFilter(MetFlowData, CI = 0.95, subject.outlier.filter = TRUE,
  path = ".")
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{CI}{Confidence interva.}

\item{subject.outlier.filter}{Remove outliers or not.}

\item{path}{Work directory.}
}
\value{
MetFlowData whose subject outliers have been removed.
}
\description{
Using PCA to filter subject outliers.
}
\examples{
\donttest{
#load the demo data
data(data, package = "metflowR")
data(sample.information, package = "metflowR")

##create a folder for demo
dir.create("demo")
setwd("demo")

# export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample.information, "sample.information.csv", row.names = FALSE)

# metflowR process
metflowR(#ImportData para
data = "data.csv",
sample.information = "sample.information.csv",
polarity = "positive",
#DataNormalization
method = "svr",
threads = 2)
## run
new.met.data <- SubjectOutlierFilter(met.data.after.pre,
                                    path = "Demo for SubjectOutlierFilter")
}
}
\seealso{
\code{\link{QCOutlierFilter}}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
