% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{SXTpaste}
\alias{SXTpaste}
\title{SXTpaste}
\usage{
SXTpaste(x, sep = " ")
}
\arguments{
\item{x}{vector.}

\item{sep}{seperate.}
}
\value{
x after paste
}
\description{
Paste for sequence.
}
\examples{
x <- c("a", "b", "c")
SXTpaste(x, sep = "")
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
