% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{SXTscale}
\alias{SXTscale}
\title{SXTscale}
\usage{
SXTscale(x, method = c("pareto", "auto"))
}
\arguments{
\item{x}{data for scale.}

\item{method}{scale method.}
}
\value{
Return x after scaling.
}
\description{
Scale method.
}
\examples{
x <- matrix(rnorm(1000),nrow = 10,ncol = 100)
x1 <- SXTscale(x)
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
