% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{SXTvip}
\alias{SXTvip}
\title{SXTvip}
\usage{
SXTvip(object)
}
\arguments{
\item{object}{PLS object.}
}
\value{
Return VIP.
}
\description{
Get VIP from pls object.
}
\examples{
library(pls)
x <- matrix(rnorm(1000),nrow = 10,ncol = 100)
y <- rep(0:1,5)
res <- plsr(y~x, method = "oscorespls")
SXTvip(res)
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
