% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{UnivariateTest}
\alias{UnivariateTest}
\title{UnivariateTest}
\usage{
UnivariateTest(MetFlowData, test.method = "t", adjust.method = "fdr",
  log.scale = FALSE, class = c("control", "case"))
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{test.method}{Which test you want to use? "t" means stutent t test
and "wilcox" mean wilcoxon test.}

\item{adjust.method}{p value correction method. See p.adjust function.}

\item{log.scale}{Data transformation method, defaulst is FALSE.}

\item{class}{Class used to do test.}
}
\value{
MetFlowData which has been added p and AUC information in tags.
}
\description{
Calculate p value and AUC for each feature.
}
\examples{
\donttest{
#load the demo data
data(data, package = "metflowR")
data(sample.information, package = "metflowR")

##create a folder for demo
dir.create("demo")
setwd("demo")

# export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample.information, "sample.information.csv", row.names = FALSE)

# metflowR process
metflowR(#ImportData para
data = "data.csv",
sample.information = "sample.information.csv",
polarity = "positive",
#DataNormalization
method = "svr",
threads = 2)
## run
new.met.data <- UnivariateTest(met.data.after.pre)
}
}
\seealso{
The details of univariate test can
be found in \code{\link[stats]{t.test}},
\code{\link[stats]{p.adjust}} and \code{\link[stats]{wilcox.test}}.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
