% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{VolcanoPlot}
\alias{VolcanoPlot}
\title{VolcanoPlot}
\usage{
VolcanoPlot(MetFlowData, x = "foldchange", y = "p", z = "vip",
  col = c("grey", "tomato"), foldchange.cutoff = c(4/3, 3/4),
  p.cutoff = 0.05, vip.cutoff = 1, path = ".")
}
\arguments{
\item{MetFlowData}{MetFlowData.}

\item{x}{x axis factor.}

\item{y}{y axis factor.}

\item{z}{z axis factor.}

\item{col}{Colour for markers and non-markers.}

\item{foldchange.cutoff}{Fold chagne cutoff.}

\item{p.cutoff}{p value cutoff.}

\item{vip.cutoff}{VIP value cutoff.}

\item{path}{The directory you want to write results.}
}
\value{
volcano plot.
}
\description{
Draw volcano plot.
}
\examples{
\donttest{
#load the demo data
data(data, package = "metflowR")
data(sample.information, package = "metflowR")

##create a folder for demo
dir.create("demo")
setwd("demo")

# export the demo data as csv
write.csv(data, "data.csv", row.names = FALSE)
write.csv(sample.information, "sample.information.csv", row.names = FALSE)

# metflowR process
metflowR(#ImportData para
data = "data.csv",
sample.information = "sample.information.csv",
polarity = "positive",
#DataNormalization
method = "svr",
threads = 2)
# FoldChange
met.data <- FoldChange(MetFlowData = met.data.after.pre,
                      to = c("1", "0"))
##PLSanalysis
PLSanalysis(met.data,
           plsmethod = "plsr",
           path = "Demo for PLSanalysis")

load("Demo for PLSanalysis/vip")
vip <- apply(vip, 2, mean)
##VIP
tags <- met.data[["tags"]]
tags <- data.frame(tags, vip)
met.data[["tags"]] <- tags
##UnivariateTest
met.data <- UnivariateTest(met.data)
## run
new.met.data <- MarkerSelection(met.data)
run
VolcanoPlot(new.met.data,
           path = "Demo for VolcanoPlot")
}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
